Move ExecuteElevated to Utilities class
