+++ targets/netcore/nanoFramework.nanoCLR.CLI/VirtualSerialBrige/VirtualSerialDeviceCommandProcessor.cs
@@ -5,10 +5,8 @@
 
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
 using System.Runtime.Versioning;
-using System.Security.Principal;
 
 namespace nanoFramework.nanoCLR.CLI
 {
@@ -93,7 +91,7 @@ namespace nanoFramework.nanoCLR.CLI
                 throw new CLIException(ExitCode.E1003);
             }
 
-            ExecuteElevated(() => VirtualSerialDeviceManager.DeleteVirtualBridge(bridge), $"{VirtualSerialDeviceManager.VirtualSerialDeviceVerb} --{VirtualSerialDeviceManager.DeleteOption} {bridge}");
+            Utilities.ExecuteElevated(() => VirtualSerialDeviceManager.DeleteVirtualBridge(bridge), $"{VirtualSerialDeviceManager.VirtualSerialDeviceVerb} --{VirtualSerialDeviceManager.DeleteOption} {bridge}");
 
             bridge = virtualComManager.GetVirtualBridgeContainingPort(port);
 
@@ -228,78 +226,6 @@ namespace nanoFramework.nanoCLR.CLI
                 Console.ForegroundColor = ConsoleColor.Yellow;
                 Console.WriteLine("No virtual serial port pairs found.");
                 Console.ForegroundColor = ConsoleColor.White;
-            }
-        }
-
-        [SupportedOSPlatform("windows")]
-        private static void ExecuteElevated(
-            Action action,
-            string arguments)
-        {
-            // check if the process is running with elevated permissions
-            bool isAdministrator = new WindowsPrincipal(WindowsIdentity.GetCurrent())
-                .IsInRole(WindowsBuiltInRole.Administrator);
-
-            if (isAdministrator)
-            {
-                action();
-            }
-            else
-            {
-                if (Program.VerbosityLevel >= VerbosityLevel.Normal)
-                {
-                    // show explanation message to user
-                    Console.ForegroundColor = ConsoleColor.Yellow;
-
-                    Console.WriteLine("");
-                    Console.WriteLine("**************************************************");
-                    Console.WriteLine("*** Requires running with elevated permissions ***");
-                    Console.WriteLine("*** UAC prompt will be shown, please authorize ***");
-                    Console.WriteLine("**************************************************");
-                    Console.WriteLine("");
-
-                    Console.ForegroundColor = ConsoleColor.White;
-                }
-
-                RunAsAdministrator(arguments);
-            }
-        }
-
-        private static void RunAsAdministrator(string arguments)
-        {
-            var fileName = Process.GetCurrentProcess().MainModule.FileName;
-
-            // this ProcessStartInfo is required to run the process with elevated permisssions
-            // can't use other options, can't route output, can't show window
-            var info = new ProcessStartInfo(fileName, arguments)
-            {
-                UseShellExecute = true,
-                WindowStyle = ProcessWindowStyle.Hidden,
-                Verb = "runas"
-            };
-
-            var process = new Process
-            {
-                EnableRaisingEvents = true,
-                StartInfo = info
-            };
-
-            process.Start();
-            
-            if(process.WaitForExit(20000))
-            {
-                // check exit code
-                var exitCode = process.ExitCode;
-
-                if(exitCode != 0)
-                {
-                    throw new CLIException(ExitCode.E9002, $"Exit code was:{exitCode}");
-                }
-            }
-            else
-            {
-                // timeout occurred, report 
-                throw new CLIException(ExitCode.E9003);
             }
         }
     }
